+++ src/states_screens/race_gui_overworld.cpp
@@ -337,7 +337,8 @@ void RaceGUIOverworld::drawGlobalMiniMap()
     }
     else if (new_rtt_mini_map != NULL)
     {
-        core::rect<s32> source(0, 0, new_rtt_mini_map->getWidth(), new_rtt_mini_map->getHeight());
+        core::rect<s32> source(0, 0, (int)new_rtt_mini_map->getWidth(),
+                               (int)new_rtt_mini_map->getHeight());
         draw2DImageFromRTT(new_rtt_mini_map->getRTT()[0],
             new_rtt_mini_map->getWidth(), new_rtt_mini_map->getHeight(),
             dest, source, NULL, video::SColor(127, 255, 255, 255), true);
